--- app/src/main/java/rooksoto/c4q/nyc/c4qhackathonapp/fragments/EmergencyFragment.java
@@ -1,24 +1,0 @@
-package rooksoto.c4q.nyc.c4qhackathonapp.fragments;
-
-import android.app.Fragment;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-/**
- * Created by rook on 2/18/17.
- */
-
-public class EmergencyFragment extends Fragment {
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return super.onCreateView(inflater, container, savedInstanceState);
-    }
-
-    @Override
-    public void onViewCreated(View view, Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-    }
-}
