+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    
+
     <string name="nyu_clinic_address">16 East 16th Street, New York, NY 10003</string>
     <string name="weill_cornell_address">505 East 70th Street, 1st Floor, New York, NY 10021</string>
     <string name="mount_sinai_address">17 East 102nd Street, 7th Floor, New York, NY 10029</string>
